Converge CLI onto plugin profilers; delete cli/lib/profilers/*
The CLI kept a parallel profiler home: cli/lib/profilers/ held its own ClickHouse
profiler, a duplicate DuckDB profiler, a Salesforce positional→options shim, and
an Elasticsearch re-export — alongside the plugin profilers it already consumed
directly (snowflake/duckdb/bigquery). That was the third profiling path (#3627).

Delete cli/lib/profilers/* entirely and point every CLI consumer (init, diff,
bin/atlas re-exports, the ES test) at each plugin's src/profiler module — the ONE
profiler home per type that MCP and the wizard also resolve (ADR-0017 / #3670).
ClickHouse + Salesforce call sites move from the positional signature to the
plugins' options shape (matching snowflake/duckdb/bigquery). Native pg/mysql stay
core via @atlas/api/lib/profiler.

CSV/Parquet ingestion (ingestIntoDuckDB) is a CLI write path, not profiling —
relocated to lib/duckdb-ingest.ts so no profiler home lingers in the CLI. Updates
the two DuckDB integration tests to the options shape.

Absorbs #3627 (+ #3665 residue). No behavior change to CLI output for any dbType.
