+++ packages/cli/src/commands/diff.ts
@@ -36,7 +36,7 @@ import {
 import {
   profileElasticsearch,
   elasticsearchConfigFromEnv,
-} from "../../lib/profilers/elasticsearch";
+} from "../../../../plugins/elasticsearch/src/profiler";
 
 export async function handleDiff(args: string[]): Promise<void> {
   // An Elastic Cloud ID names the endpoint without a URL (#3309). Trimmed so
@@ -203,8 +203,8 @@ export async function handleDiff(args: string[]): Promise<void> {
           );
           break;
         case "clickhouse": {
-          const { profileClickHouse } = await import("../../lib/profilers/clickhouse");
-          result = await profileClickHouse(connStr, filterTables);
+          const { profileClickHouse } = await import("../../../../plugins/clickhouse/src/profiler");
+          result = await profileClickHouse({ url: connStr, selectedTables: filterTables });
           break;
         }
         case "snowflake": {
@@ -226,8 +226,8 @@ export async function handleDiff(args: string[]): Promise<void> {
           break;
         }
         case "salesforce": {
-          const { profileSalesforce } = await import("../../lib/profilers/salesforce");
-          result = await profileSalesforce(connStr, filterTables);
+          const { profileSalesforce } = await import("../../../../plugins/salesforce/src/profiler");
+          result = await profileSalesforce({ url: connStr, selectedTables: filterTables });
           break;
         }
         case "bigquery": {
